requirements.txt
asgiref==3.4.1
django-taggit==1.5.1
django==3.0.14
pytz==2021.1
sqlparse==0.4.1
